Cast as float before KMeans
Temporary hot fix for a bug in PySpark (it should automatically upcast
to float before running KMeans, but is currently failing to do so,
leading to a SerDe error)

+++ python/thunder/clustering/kmeans.py
@@ -130,7 +130,7 @@ class KMeans(object):
 
         import pyspark.mllib.clustering as mllib
 
-        model = mllib.KMeans.train(data.rdd.values(), k=self.k, maxIterations=self.maxIterations)
+        model = mllib.KMeans.train(data.astype('float').rdd.values(), k=self.k, maxIterations=self.maxIterations)
 
         return KMeansModel(asarray(model.clusterCenters))
